refactor: moved enableInContext value from useContext to redux
